+++ src/components/DifficultyLabel/constant.ts
