+++ src/components/DifficultyLabel/index.tsx
